requirements.txt
keras==2.6.0
matplotlib==3.4.3
numpy==1.19.5
Pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
pytorch-lightning==1.4.9
